Split out the ssp definitions
I was looking to offer another ssp definition but I was worried the main
file was getting a bit too large, so I split the existing definitions
out first.
