+++ .changeset/tasty-donkeys-scream.md
@@ -1,0 +1,5 @@
+---
+'sveltekit-search-params': minor
+---
+
+Split out the ssp definitions
