+++ src/lib/index.ts
@@ -1,2 +1,5 @@
 // Reexport your entry components here
-export { ssp, queryParameters, queryParam } from './sveltekit-search-params.js';
+import ssp from './ssp';
+
+export { queryParameters, queryParam } from './sveltekit-search-params';
+export { ssp };
