+++ src/lib/ssp/index.ts
@@ -1,0 +1,24 @@
+import {
+	arrayEncodeAndDecodeOptions,
+	objectEncodeAndDecodeOptions,
+} from './json-encode';
+import { lzEncodeAndDecodeOptions } from './lz-encode';
+import { primitiveEncodeAndDecodeOptions } from './util';
+
+export default {
+	number: primitiveEncodeAndDecodeOptions({
+		encode: (value: number) => value.toString(),
+		decode: (value: string | null) => (value ? parseFloat(value) : null),
+	}),
+	boolean: primitiveEncodeAndDecodeOptions({
+		encode: (value: boolean) => value + '',
+		decode: (value: string | null) => value !== null && value !== 'false',
+	}),
+	string: primitiveEncodeAndDecodeOptions({
+		encode: (value: string | null) => value ?? '',
+		decode: (value: string | null) => value,
+	}),
+	object: objectEncodeAndDecodeOptions,
+	array: arrayEncodeAndDecodeOptions,
+	lz: lzEncodeAndDecodeOptions,
+} as const;
